fix(ssm): preserve dtype through state-space scan and support pytree autotune outputs
- Fix mixed-precision dtype promotion in SSM forward/backward passes by
  explicitly casting state, discretized matrices, and output to preserve
  the caller's dtype (e.g. bf16 state with fp32 A)
- Handle tuple/pytree kernel outputs in autotune backward validation
  by summing all numeric leaves instead of assuming a single array
- Bump version to 0.0.66

+++ ejkernel/kernels/_xla/state_space_v1/_interface.py
@@ -145,7 +145,7 @@ def _ssm1_fwd_rule(
     if initial_state is None:
         initial_state = jnp.zeros(
             (batch_size, intermediate_size, ssm_state_size),
-            dtype=jnp.float32,
+            dtype=hidden_states.dtype,
         )
 
     if use_single_step and seq_len == 1:
